add some command and OS guardrails

+++ bash_profile
@@ -41,9 +41,9 @@ export NVM_DIR="$HOME/.nvm"
 # Safe Python version management using pyenv
 # https://opensource.com/article/19/5/python-3-default-mac
 if command -v pyenv 1>/dev/null 2>&1; then
-  eval "$(pyenv init -)"
+    PATH="$(pyenv root)/shims:$PATH"
+    eval "$(pyenv init -)"
 fi
-PATH="$(pyenv root)/shims:$PATH"
 
 # Paths
 export PATH="/usr/local/sbin:$PATH"
@@ -72,14 +72,16 @@ alias mk=minikube
 export BASH_COMPLETION_COMPAT_DIR=/usr/local/etc/bash_completion.d
 # shellcheck disable=SC1091
 [[ -r "/usr/local/etc/profile.d/bash_completion.sh" ]] && . "/usr/local/etc/profile.d/bash_completion.sh"
-# shellcheck disable=SC1091
-source /Users/bowena/Documents/Projects/VMware/DX/cli_taxo/exp4/results/velero_completion.sh
+if command -v kubectl 1>/dev/null 2>&1; then
+    # shellcheck disable=SC1090
+    source <(kubectl completion bash)
+    complete -F __start_kubectl k
+fi
+if command -v minikube 1>/dev/null 2>&1; then
 # shellcheck disable=SC1090
-source <(kubectl completion bash)
-complete -F __start_kubectl k
-# shellcheck disable=SC1090
-source <(minikube completion bash)
-complete -F __start_minikube mk
+    source <(minikube completion bash)
+    complete -F __start_minikube mk
+fi
 
 # Functions
 
@@ -95,31 +97,20 @@ kubeconf() {
 }
 kubeconf >/dev/null
 
-brew_update() {
-    # Additional homebrew housekeeping
-    brew update && brew upgrade && brew cleanup; 
-}
-# Automate homebrew update
-# TODO: Check for OS before applying Darwin specific stuff
-UPDATE_DATE="$HOME/.last_update"
-[ -f "$UPDATE_DATE" ] || echo "00" > "$UPDATE_DATE"
-if [ "$CDATE" != "$(head -n 1 "$UPDATE_DATE")" ]; then
-    echo "$CDATE" > "$UPDATE_DATE"
-    # shellcheck disable=SC2154
-    echo -e "Checking homebrew..."
-    brew_update
-fi
-
-vmw_whois() {
-    # VMware specific whois
-    # TODO: Migrate to seperate tool under vmware scripts
-    local name="$*"
-    #ref https://source.vmware.com/portal/search/people?q=alister&aq=(@cnbd%3D%22alister%22%20OR%20@ucnbd%3D%22alister%22)&client=InternalPeopleSearch&Tab=vmwarepeople&start=0&num=20&sid=1606940050&allPeople=true
-    local url_base='https://source.vmware.com/portal/search/people?'
-    local url_query_attributes="client=InternalPeopleSearch&Tab=vmwarepeople&start=0&num=20&sid=1606938064&allPeople=true"
-    name="${name//+([[:space:]])/%20}"
-    local url_query="q=${name}&aq=(@cnbd%3D%22${name}%22%20OR%20@ucnbd%3D%22${name}%22)"
-    open "${url_base}${url_query}&${url_query_attributes}"
+[[ "$OSTYPE" == 'darwin'* ]] && {
+    brew_update() {
+        # Additional homebrew housekeeping
+        brew update && brew upgrade && brew cleanup; 
+    }
+    # Automate homebrew update
+    UPDATE_DATE="$HOME/.last_update"
+    [ -f "$UPDATE_DATE" ] || echo "00" > "$UPDATE_DATE"
+    if [ "$CDATE" != "$(head -n 1 "$UPDATE_DATE")" ]; then
+        echo "$CDATE" > "$UPDATE_DATE"
+        # shellcheck disable=SC2154
+        echo -e "Checking homebrew..."
+        brew_update
+    fi
 }
 
 # Prompts
@@ -148,18 +139,20 @@ PS4="${cyan}$0.$LINENO ⨠${normal} " # tracing
 
 # History manager
 # @ref https://github.com/dvorka/hstr/blob/master/CONFIGURATION.md
-alias hh=hstr                    # hh to be alias for hstr
-export HSTR_CONFIG=hicolor       # get more colors
-shopt -s histappend              # append new history items to .bash_history
-export HISTCONTROL=ignorespace   # leading space hides commands from history
-export HISTFILESIZE=10000        # increase history file size (default is 500)
-export HISTSIZE=${HISTFILESIZE}  # increase history size (default is 500)
-# ensure synchronization between bash memory and history file
-export PROMPT_COMMAND="history -a; history -n; ${PROMPT_COMMAND}"
-# if this is interactive shell, then bind hstr to Ctrl-r (for Vi mode check doc)
-if [[ $- =~ .*i.* ]]; then bind '"\C-r": "\C-a hstr -- \C-j"'; fi
-# if this is interactive shell, then bind 'kill last command' to Ctrl-x k
-if [[ $- =~ .*i.* ]]; then bind '"\C-xk": "\C-a hstr -k \C-j"'; fi
+if command -v hstr 1>/dev/null 2>&1; then
+    alias hh=hstr                    # hh to be alias for hstr
+    export HSTR_CONFIG=hicolor       # get more colors
+    shopt -s histappend              # append new history items to .bash_history
+    export HISTCONTROL=ignorespace   # leading space hides commands from history
+    export HISTFILESIZE=10000        # increase history file size (default is 500)
+    export HISTSIZE=${HISTFILESIZE}  # increase history size (default is 500)
+    # ensure synchronization between bash memory and history file
+    export PROMPT_COMMAND="history -a; history -n; ${PROMPT_COMMAND}"
+    # if this is interactive shell, then bind hstr to Ctrl-r (for Vi mode check doc)
+    if [[ $- =~ .*i.* ]]; then bind '"\C-r": "\C-a hstr -- \C-j"'; fi
+    # if this is interactive shell, then bind 'kill last command' to Ctrl-x k
+    if [[ $- =~ .*i.* ]]; then bind '"\C-xk": "\C-a hstr -k \C-j"'; fi
+fi
 
 # Additional configurations/overrides
 # shellcheck disable=SC1091
